weather-agent/pyproject.toml
[project]
name = "weather-agent"
version = "0.1.0"
description = "A LangChain agent that uses the weather MCP server"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "mcp>=1.6.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/weather_agent"] 